\name{CorporaCoCo-package}
\alias{CorporaCoCo-package}
\alias{CorporaCoCo}
\docType{package}

\title{Comparing Co-occurrence between corpora.}
\description{
  Implements the method described in Hennessey and Wiegand et al. (2017).
}
\details{
  A good place to start is the
  \sQuote{\href{../doc/proof_of_concept.pdf}{Proof of Concept}} vignette.
  There is also a \sQuote{\href{../doc/faq.pdf}{FAQ}} vignette.
  You can see a list of package vignettes with
  \code{vignette(package = "CorporaCoCo")} and you can see a particular
  vignette with something like
  \code{vignette("faq", package = "CorporaCoCo")}.

  For a list of all documentation use \code{library(help="CorporaCoCo")}.
}

\author{
  Maintainer: Michaela Mahlberg <m.j.allaway@bham.ac.uk>.
}

\references{
  A. Hennessey and V. Wiegand and C. R. Tench and M. Mahlberg (2017)
  \emph{Comparing co-occurrences between corpora.} In preparation.
}
