\name{CorporaCoCo-package}
\alias{CorporaCoCo-package}
\alias{CorporaCoCo}
\docType{package}

\title{Comparing Co-occurrence between corpora.}
\description{
  Implements the method described in Hennessey and Wiegand et al. (2017).
}
\details{
  A good place to start is the
  \sQuote{\href{../doc/intro.html}{Introduction to CorporaCoCo}} vignette.
  You can open the vignette with
  \code{vignette("intro", package = "CorporaCoCo")}.

  For a list of all documentation use \code{library(help="CorporaCoCo")}.
}

\author{
  Maintainer: Anthony Hennessey <anthony.hennessey@nottingham.ac.uk>.
}

\references{
  A. Hennessey and V. Wiegand and C. R. Tench and M. Mahlberg (2017)
  \emph{Comparing co-occurrences between corpora.} In preparation.
}
